% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetics.R
\name{add}
\alias{add}
\title{Add two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
sum of two numbers: \code{x} and \code{y}
}
\description{
The function works with numbers and does arithmrtic addition
}
