% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetics.R
\name{add2num}
\alias{add2num}
\title{Add two numbers}
\usage{
add2num(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{A number}
}
\value{
sum of two numbers: \code{x} and \code{y}
}
\description{
The function works with numbers and does arithmrtic addition
}
\examples{
add2num(1, 2)
add2num(-7, 0)
}
