% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arithmetics.R
\name{negate}
\alias{negate}
\title{negate number}
\usage{
negate(x)
}
\arguments{
\item{x}{A number}
}
\value{
a number with opposite sign
}
\description{
The function negates number passed as a parameter.
  \code{negate(1)} returns \code{-1}
}
\examples{
negate(1)
negate(-1)
}
