% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simplify.rules}
\alias{simplify.rules}
\title{Rules for names simplification}
\format{A dataframe with 4 variables \describe{ \item{Seq}{rule sequence, the
  less number the earlier applied} \item{From}{string to find (pattern in
  terms of gsub)} \item{To}{replacement} \item{Fixed}{pattern search mode: if
  TRUE pattern is a string to be matched as is. Overrides all conflicting
  arguments} }}
\usage{
simplify.rules
}
\description{
A dataset contains rules that a subsequently applied to text string. Each
rule simplifes text using standard \code{gsub} function. Order of rules in
dataset matters - first rules apply earlier
}
\keyword{datasets}
