% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{package}
\name{hello.rpkg}
\alias{hello.rpkg}
\alias{hello.rpkg-package}
\title{hello.rpkg: A sandbox package for learning purposes}
\description{
The package includes content for successful build. It is used to learn ideas
behind package creaion and get practice.
}
\details{
Documentation allows to use styles

\emph{empth}, \strong{strong}. It is possible to include links to functions
\code{\link{add2num}} and to internet resources
\href{http://rstudio.com}{Rstudio} or to emails \email{name@host.com}
}
